refactor(keypress): process Tab key press directly.
No need to subscribe focus event any more. This prevents handle click
event twice, one for `click` event, and one for `focus` event. And there
is a issue on focusing on Firefox when poping up dialog box like `alert`
or `prompt`, when it is closed, Firefox restores focus to previous item.